pyproject.toml
[project]
name = "umap-parameters"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "seaborn>=0.13.2",
    "umap-learn>=0.5.11",
]
